Do not specify count for data source aws_ami.mod
If the ami being passed into this module comes as a computed value (for example, because we’re making an `aws_ami_copy` from another image), the count value becomes a computed value, and Terraform does not allow this (so it results in an error).

Since a data source doesn’t create any resources, we don’t need to specify its count—it can always be present, and in the resource `aws_instance` we’ll just use it, or not.