Remove ability to specify a different region in //aws/database_replica
Having to specify providers in the module (so we can use a different region) caused problems when trying to delete a replica — Terraform’s explicit provider support requires that all providers be specified explicitly when they are configured in a module, so uses of this module that did not do that could not be deleted.

Since this feature is not used anywhere anymore, per Lloyd, it’s easiest to remove it.