Spin up CloudWatch for CloudTrail logs
As a first step toward improving our AWS security defaults, we'd like
to enable CloudWatch any time we're spinning up a CloudTrail trail.
This will give us a lot of freedom moving forward to set up alerts,
monitoring, et cetera.

As part of this change, we're also making a few tweaks to our defaults
for the `cloudtrail` module:

- Enable logging for the CloudTrail log bucket
- Enable log file validation for CloudTrail logs
- Use an `aws_iam_policy_document` data source for our bucket policy

There are a few best practices we're _not_ going to try to implement
right now -- encrypting our log bucket with a KMS key, for instance.
But this should be a good step in the right direction to make security
a first-class citizen of our infrastructure management.