Pull CloudWatch into its own terraform module
While there are some concerns that we don't necessarily want to add
CloudWatch for every account -- and I stand by the initial concept
there -- there's a much larger concern about the composability of the
current approach, which doesn't really allow for resources to actually
be added on top of CloudWatch for purposes of alerting, etc.

To address this larger concern, we're extracting the creation of
CloudWatch resources into a separate module, which gives us the Role
and Group ARNs (the only two meaningful pieces of information for
CloudTrail) as outputs, and allows for customization of the list of
services that CloudWatch is allowed to collect information about.