Change NAT gateway module to create a separate subnet for each AZ.
These subnets will be used by the k8s LoadBalancer for its ingress targets,
because we want to put the k8s nodes themselves in private subnets, which route
to the NAT gateway for egress.

This also adjusts the syntax for using aws_route, since there is 'gateway_id'
and also 'nat_gateway_id', so we use the proper attribute at the proper time.

I was not able to use the vpc/subnets module to do the NAT-gateway-subnets creation
because you cannot use 'count' arguments with 'modules', and thus using the
'uses_nat_gateway' option to enable/disable the nat-gateway creation requires
that that module NOT depend on other modules.